pyproject.toml
[tool.poetry]
name = "django-perfumes-app"
version = "0.1.0"
description = ""
authors = ["Claudia <claudiaescola7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
django-environ = "^0.8.1"
python-dotenv = "^0.20.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
dj-rest-auth = "^2.2.4"
PyJWT = "^2.3.0"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.950"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
